+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - [matterbridge]: Added a check of the current Matterbridge version (required v1.5.5).
+- [plugin]: Added configuration of ColorControl cluster features (HS, XY, CT)
 
 <a href="https://www.buymeacoffee.com/luligugithub">
   <img src="./yellow-button.png" alt="Buy me a coffee" width="120">